+++ ui/dialogaddexpense.cpp
@@ -14,7 +14,7 @@ DialogAddExpense::DialogAddExpense(QWidget *parent) :
     ui->dateEdit->setDate(QDate::currentDate());
 
     ui->dateEdit->setMaximumDate(QDate::currentDate());
-    ui->lineEditMoneyValue->setValidator( new QDoubleValidator(0.0, std::numeric_limits<float>::max(), 2, this));
+    ui->lineEditMoneyValue->setValidator( new QDoubleValidator(0.0, std::numeric_limits<double>::max(), 2, this));
 
     date = QDate::currentDate();
     moneyValue = 0;
@@ -65,7 +65,7 @@ void DialogAddExpense::on_dateEdit_userDateChanged(const QDate &date)
 
 void DialogAddExpense::on_lineEditMoneyValue_textChanged(const QString &arg1)
 {
-    this->moneyValue = arg1.toFloat();
+    this->moneyValue = arg1.toDouble();
 }
 
 void DialogAddExpense::on_comboBoxCurrency_currentTextChanged(const QString &arg1)
